optimize english mode by memoization
